kernel: replace unsafe uses of 1L
On some 64bit systems (most notably, Windows), sizeof(long) == 4, and so
expressions like `1L << 60` overflow. Moreover, an enum strictly
speaking is only guaranteed to be able to represent an int, and so using
it to represent 64bit constants is not safe; we hence turn
INT_INTOBJ_MIN and INT_INTOBJ_MAX into global static constants.

+++ src/bits_intern.h
@@ -34,7 +34,7 @@
 /* constructs a mask that selects bits <from> to <to> inclusive of a UInt */
 static inline UInt MaskForCopyBits(UInt from, UInt to)
 {
-    return ((to == BIPEB - 1) ? 0 : (1L << (to + 1))) - (1L << from);
+    return ((to == BIPEB - 1) ? 0 : ((UInt)1 << (to + 1))) - ((UInt)1 << from);
 }
 
 /* copies a block of bits from the UInt <from> to the one pointed at
